Update on "[dtensor] move tensor constructors to a separate module"
This is to ensure __init__.py only contain public APIs

cc XilunWu H-Huang awgu kwen2501 fegin fduwjj wz337 wconstab d4l3k c-p-i-o

[ghstack-poisoned]
