+++ torch/distributed/_tensor/ops/utils.py
@@ -113,25 +113,6 @@ def normalize_dims(dims: Union[int, Sequence[int]], ndim: int) -> Sequence[int]:
     return dims
 
 
-def normalize_to_torch_size(size) -> torch.Size:
-    """
-    Unify variable types of size argument to torch.Size
-    Acceptable types include:
-        int, Sequence[int], Tuple[int], Tuple[Sequence[int]],
-        or torch.Size
-    """
-    if isinstance(size, torch.Size):
-        return size
-
-    if isinstance(size, int):
-        torch_size = [size]
-    elif len(size) == 1 and isinstance(size[0], Sequence):
-        torch_size = list(size[0])
-    else:
-        torch_size = list(size)
-    return torch.Size(torch_size)
-
-
 def prod(xs: Iterable[int]) -> int:
     return functools.reduce(operator.mul, xs, 1)
